Day 12: Inheritance
You are given two classes, Person and Student, where Person is the base class and Student is the derived class. Completed code for Person and a declaration for Student are provided for you in the editor. Observe that Student inherits all the properties of Person.
Complete the Student class by writing the following:

A Student class constructor, which has  4  parameters.
And a calculate() function which returns the grade of student.